+++ analysis/test_menu_children.lua
@@ -23,7 +23,11 @@ package.preload["./scripts/actionSequenceEditor"] = function()
 end
 package.preload["./scripts/actionSequenceRunner"] = function()
 	return dofile("actionSequenceRunner.lua")
+	end
+package.preload["./scripts/position"] = function()
+	return dofile("position.lua")
 end
+
 
 training_settings = {}
 globals = {}